pyproject.toml
[project]
name = "django-rest-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.8.1",
    "django==5.1.1",
    "django-extensions==3.2.3",
    "django-filter>=25.1",
    "django-silk>=5.4.0",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "markdown>=3.8",
    "pillow==10.4.0",
    "pydot>=4.0.0",
    "pyparsing>=3.2.3",
    "python-dotenv>=1.1.0",
    "sqlparse==0.5.1",
    "tzdata==2024.2",
]
